requirements.txt
Flask==2.0.1
gunicorn==20.1.0
flask-cors
flask_api
PyJWT
firebase_admin